경북대 BE_배준호 3주차 과제(2단계) (#322)
* feat : step 0, 기본 코드 설정

* docs : step 1 요구 사항 작성

* build : 빌드 파일 및 property 파일 작성

JPA에 대한 설정 추가

* feat : Entity 구현

기존 domain 클래스를 Entity 어노테이션을 통해 DB와 직접 매핑되는 Entity 클래스로 만들었다

* refactor : JPA 사용 Repository 구현

JpaRepository를 상속받는 interface 방식으로 repository 클래스를 변경하고, 이를 사용하는 service 클래스 역시 이에 맞게 바꾸었다

* fix : 더 이상 사용하지 않는 전처리 코드 삭제

* refactor : front 부분에서도 변화 적용

wish 객체에 수량을 추가한 부분, product 정보가 아닌 wishProduct 정보 전송하는 부분

* feat : 수량 수정 기능 추가

피드백에 따라 2주차 과제의 미구현 요구 사항 구현

* test : jpaDataTest 구현

* refactor : 피드백 사항 적용

* feat : separate user page and admin page

* chore : 더 이상 사용하지 않는 import 문 삭제

* docs : 새로운 요구사항 반영

* style : DTO 클래스를 Response와 Request로 분리

* fix : 삭제된 파일

* refactor : change wishList's field to Other Entities

~id 형식이었던 wishList 클래스 필드 변수들을 실제 객체로 변경

* feat : 관리자 페이지와 유저 페이지 분리 및 위시리스트 페이지 구현

* style : 접근 제한자 지정

이전에 왜 접근 제한자를 지웠지? 아무튼 private 접근 제한자를 통해 외부에서 클래스 필드 변수 접근 못하게 바꾸었다

* fix : Transactional 어노테이션 삽입

CUD 동작에 대해 트랜잭션 수행되게 함 ( 안하면 update나 delete 시 오류가 나거나 바로 반영안되고 쿼리 캐싱됨 )

* chore : 메서드 명 변경

product 쪽에 대한 반영

* chore : 메서드 매개 변수를 userId에서 id로 변경

이 편이 좀 더 직관적인 것 같아서 변경

* chore : 백에서의 변화 반영

* style : change return value

userResponse를 반환하여 id를 사용할 수 있게 변경

* feat : implement delete logic

받는 값을 id가 포함된 userRes로 변경하고, 이를 통해 로직 간소화, 그리고 delete 로직에서 수량을 고려하도록 변경

---------

+++ src/main/java/gift/service/WishListService.java
@@ -1,60 +1,99 @@
 package gift.service;
 
-import gift.DTO.WishProductDTO;
+import gift.DTO.Product.ProductRequest;
+import gift.DTO.Product.ProductResponse;
+import gift.DTO.User.UserRequest;
+import gift.DTO.User.UserResponse;
+import gift.DTO.Wish.WishProductRequest;
+import gift.DTO.Wish.WishProductResponse;
+import gift.domain.Product;
+import gift.domain.User;
 import gift.domain.WishProduct;
+import gift.repository.ProductRepository;
+import gift.repository.UserRepository;
 import gift.repository.WishListRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class WishListService {
     private final WishListRepository wishListRepository;
+    private final ProductRepository productRepository;
+    private final UserRepository userRepository;
 
-    public WishListService(WishListRepository wishListRepository){
+    public WishListService(
+            WishListRepository wishListRepository, ProductRepository productRepository, UserRepository userRepository
+    ){
         this.wishListRepository = wishListRepository;
-    }
-    /*
-     * 특정 상품을 위시리스트에 추가하는 로직
-     */
-    public void addWishList(WishProductDTO wishProductDTO){
-        String userId = wishProductDTO.getUserId();
-        Long productId = wishProductDTO.getProductId();
-
-        if(wishListRepository.existsByUserIdAndProductId(userId, productId)){
-            WishProduct wishProduct = wishListRepository.findByUserIdAndProductId(userId, productId);
-            wishProduct.changeCount(wishProduct.getCount() + 1);
-            wishListRepository.save(wishProduct);
-            return;
-        }
-
-        WishProduct wishProduct = new WishProduct(userId, productId);
-        wishListRepository.save(wishProduct);
+        this.productRepository = productRepository;
+        this.userRepository = userRepository;
     }
     /*
      * 특정 유저의 위시리스트를 반환하는 로직
      */
-    public List<WishProductDTO> loadWishList(String userId){
-        List<WishProductDTO> list = new ArrayList<>();
+    public List<WishProductResponse> loadWishList(Long id){
+        List<WishProductResponse> list = new ArrayList<>();
 
-        List<WishProduct> wishes = wishListRepository.findByUserId(userId);
+        List<WishProduct> wishes = wishListRepository.findByUserId(id);
         for (WishProduct wishProduct : wishes) {
-            WishProductDTO wishProductDTO = new WishProductDTO(
-                    wishProduct.getUserId(),
-                    wishProduct.getProductId(),
+            User userEntity = wishProduct.getUser();
+            Product productEntity = wishProduct.getProduct();
+
+            UserResponse user = new UserResponse(
+                    userEntity.getId(),
+                    userEntity.getUserId(),
+                    userEntity.getEmail(),
+                    userEntity.getPassword()
+            );
+
+            ProductResponse product = new ProductResponse(
+                    productEntity.getId(),
+                    productEntity.getName(),
+                    productEntity.getPrice(),
+                    productEntity.getImageUrl()
+            );
+
+            WishProductResponse wishProductResponse = new WishProductResponse(
+                    wishProduct.getId(),
+                    user,
+                    product,
                     wishProduct.getCount()
             );
-            list.add(wishProductDTO);
+
+            list.add(wishProductResponse);
         }
 
         return list;
     }
     /*
+     * 특정 상품을 위시리스트에 추가하는 로직
+     */
+    @Transactional
+    public void addWishList(WishProductRequest wishProductRequest){
+        Long id = wishProductRequest.getUser().getId();
+        Long productId = wishProductRequest.getProduct().getId();
+
+        if(wishListRepository.existsByUserIdAndProductId(id, productId)){
+            WishProduct wishProduct = wishListRepository.findByUserIdAndProductId(id, productId);
+            wishProduct.changeCount(wishProduct.getCount() + 1);
+
+            return;
+        }
+
+        User byId = userRepository.findById(id).orElseThrow(NullPointerException::new);
+        Product byProductId = productRepository.findById(productId).orElseThrow(NullPointerException::new);
+        WishProduct wishProduct = new WishProduct(byId, byProductId);
+        wishListRepository.save(wishProduct);
+    }
+    /*
      * 특정 유저의 특정 위시리스트 물품의 수량을 변경하는 로직
      */
-    public void updateWishProduct(String userId, Long productId, int count){
+    @Transactional
+    public void updateWishProduct(Long userId, Long productId, int count){
         WishProduct wish = wishListRepository.findByUserIdAndProductId(userId, productId);
         wish.changeCount(count);
         wishListRepository.save(wish);
@@ -63,8 +102,13 @@ public class WishListService {
      * 특정 유저의 특정 위시리스트 물품을 삭제하는 로직
      */
     @Transactional
-    public void deleteWishProduct(String userId, Long productId){
-        wishListRepository.deleteByUserIdAndProductId(userId, productId);
+    public void deleteWishProduct(Long userId, Long productId){
+        WishProduct wish = wishListRepository.findByUserIdAndProductId(userId, productId);
+        if(wish.getCount() == 1) {
+            wishListRepository.deleteByUserIdAndProductId(userId, productId);
+            return;
+        }
+        wish.changeCount(wish.getCount() - 1);
     }
 
 }
